# apps/server/static/openapi.yaml
openapi: 3.0.3
info:
  title: BTSearch API
  description: |
    API documentation for BTSearch Backend
    Last updated: 2026-10-06
  version: 1.4.3
servers:
  - url: https://btsearch.pl/api/v1
    description: API v1 base path
...
paths:
...
  /stations/{station_id}/uplink:
    patch:
      tags: [Stations]
      summary: Set or remove station uplink
      description: "Creates or replaces the station uplink (backhaul). Send `type: null` to remove it (204, or 404 when the station has no uplink). `model` is only stored for microwave links and is cleared for other types. Requires `update:stations` permission"
      security:
        - ApiKeyAuth: []
        - CookieAuth: []
      parameters:
        - $ref: "#/components/parameters/StationIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StationUplinkPatch" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/StationUplink"
                      - type: object
                        required: [station_id]
                        properties:
                          station_id: { type: integer }
        "204": { description: Uplink removed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API Key for authentication via X-Api-Key header
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth
      description: Cookie-based authentication
  parameters:
...
    StationIdPath:
      name: station_id
      in: path
      required: true
      schema: { type: integer }
  responses:
    BadRequest:
      description: Bad request (invalid parameters, validation error, or invalid query)
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [BAD_REQUEST, VALIDATION_ERROR, INVALID_QUERY]
          example:
            errors:
              - code: BAD_REQUEST
                message: Invalid request
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [UNAUTHORIZED]
          example:
            errors:
              - code: UNAUTHORIZED
                message: Unauthorized
    Forbidden:
      description: Forbidden or insufficient permissions
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [FORBIDDEN, INSUFFICIENT_PERMISSIONS]
          example:
            errors:
              - code: FORBIDDEN
                message: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [NOT_FOUND]
          example:
            errors:
              - code: NOT_FOUND
                message: Resource not found
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [TOO_MANY_REQUESTS]
          example:
            errors:
              - code: TOO_MANY_REQUESTS
                message: Too many requests
...
    InternalServerError:
      description: Internal server error (e.g. failed to create/update/delete)
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [INTERNAL_SERVER_ERROR, FAILED_TO_CREATE, FAILED_TO_UPDATE, FAILED_TO_DELETE, AUTH_FAILURE]
          example:
            errors:
              - code: INTERNAL_SERVER_ERROR
                message: Internal server error
  schemas:
...
    ApiError:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          minItems: 1
          items:
            type: object
            required: [code, message]
            properties:
              code:
                type: string
                enum:
                  - INTERNAL_SERVER_ERROR
                  - BAD_REQUEST
                  - NOT_FOUND
                  - UNAUTHORIZED
                  - FORBIDDEN
                  - FAILED_TO_DELETE
                  - INVALID_QUERY
                  - FAILED_TO_UPDATE
                  - FAILED_TO_CREATE
                  - AUTH_FAILURE
                  - VALIDATION_ERROR
                  - ALREADY_LOGGED_IN
                  - INSUFFICIENT_PERMISSIONS
                  - TOO_MANY_REQUESTS
                  - QUOTA_EXCEEDED
                  - DUPLICATE_ENTRY
                  - CONFLICT
                  - SERVICE_UNAVAILABLE
                  - TWO_FACTOR_REQUIRED
                  - DUPLICATE_REQUEST
              message: { type: string }
              details:
                type: array
                description: Present for validation errors and endpoint-specific detailed responses
                items:
                  type: object
                  properties:
                    field: { type: string }
                    validationMessage: { type: string, nullable: true }
...
    StationUplink:
      type: object
      required: [id, type, speed, model]
      properties:
        id: { type: integer }
        type: { type: string, enum: [fiber, microwave, satellite], description: "Backhaul type: fiber, microwave radio link or satellite link" }
        speed: { type: integer, nullable: true, minimum: 1, description: "Link capacity in Mbps" }
        model: { type: string, nullable: true, maxLength: 100, description: "Microwave link equipment model. Always null for other types" }
    StationUplinkPatch:
      type: object
      required: [type]
      properties:
        type: { type: string, enum: [fiber, microwave, satellite], nullable: true, description: "Uplink type. `null` removes the uplink" }
        speed: { type: integer, nullable: true, minimum: 1, maximum: 2147483647, description: "Link capacity in Mbps" }
        model: { type: string, nullable: true, maxLength: 100, description: "Microwave link equipment model. Cleared for other types" }
...